# perf: parallelize side effects optimization