<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

Refactor our implementation to more parallel friendly and incremental friendly, and parallelize it as much as possible.

Time cost on side effects optimization:

- internal project (FS): 180ms -> 63ms
- arco-pro: 3.5ms -> 1.9ms
- 20000 modules: 2.9ms -> 2.9ms (basically no change since there are no connections are optimizable in this case)

<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).